feat: Replace the parser generated by Jison with a hand-built parser from JJU
BREAKING CHANGE: There is no `yy.parseError` to intercept error handling. Use the thrown error - it contains all available information. The error does not include the `hash` object with structured information. Look for the [documentd properties](/prantlf/jsonlint#error-handling). The location of the error occurrence is available as `location.start`, for example.

DEPRECATION: The only exposed object to use from now on is the `parse` method as a named export. Other exports (`parser` and `Parser`) are deprecated and will be removed in future.

The parser from ["Utilities to work with JSON/JSON5 documents"](/rlidwka/jju) is four times faster, than the prevous one, has approximatly the same size and can be easier enhanced, regarding both features and error handling.